face-recognition/requirements.txt
boto3
pillow
facenet-pytorch
torch
opencv-python-headless